refactor(server): three follow-up review fixes on PR 160
- defineUserAssignmentTable: derive `fkColumnName` (camelCase→snake_case) and
  `checkName` (`<tableName>_assignment_source_check`) instead of taking them
  as args. Drops 2 of 5 args; impossible now to pass mismatched key/column
  pairs. drizzle-kit generate confirms zero schema diff.
- AssignmentSource constants: move the canonical definitions and the union
  type from `userAssignmentsRepo` into `_userAssignmentTable.ts` so the
  schema CHECK/`$type<>`/`default(...)` and the repo helpers share one
  source of truth. Breaks the cycle the PR description called out (the
  schema-level helper sits below userAssignmentsRepo in the import graph,
  not above). userAssignmentsRepo re-exports them so all existing callers
  keep their import path.
- routes/clients.ts: hoist the 16-field array passed to
  parseOptionalStringFields to module scope (`PATCH_OPTIONAL_STRING_FIELDS`)
  so it isn't reallocated per request.

bun test test → 672/672 pass; tsc clean; biome clean on all 20 PR files;
db:generate → no schema changes.
